fix: better view
