+++ scripts/deploy-static.sh
